mediagent-agent/requirements.txt
langchain>=0.3.0
langgraph>=0.2.50
langchain-anthropic>=0.3.0
python-dotenv>=1.0.0
